+++ CHANGES.md
@@ -80,6 +80,10 @@
   a valid Cryptol identifier.
   ([#2036](https://github.com/GaloisInc/cryptol/issues/2036))
 
+* Add `pIsNeq` to `Cryptol.TypeCheck.Type`, which recognizes if a Cryptol
+  constraint is headed by a not-equal (`!=`) operator.
+  ([#2038](https://github.com/GaloisInc/cryptol/issues/2038))
+
 # 3.5.0 -- 2026-01-27
 
 ## Administrative changes